% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_various.R
\name{as.character.formula}
\alias{as.character.formula}
\title{Formula to character}
\usage{
\method{as.character}{formula}(x, ...)
}
\arguments{
\item{x}{formula}

\item{\ldots}{unused}
}
\description{
Formula to character
}
\examples{
f <-  aha ~ hih + oho
as.character(f)
}
