% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_cdl_cols.R
\name{as.data.frame.Extent}
\alias{as.data.frame.Extent}
\title{Extent to df}
\usage{
\method{as.data.frame}{Extent}(x, row.names, optional, ...)
}
\arguments{
\item{x}{extent of raster}

\item{row.names, optional, \ldots}{unused (but passed)}
}
\description{
Extent to df
}
