% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_various.R
\name{as.data.frame.packageIQR}
\alias{as.data.frame.packageIQR}
\title{As data frame for data()}
\usage{
\method{as.data.frame}{packageIQR}(x, ..., stringsAsFactors = FALSE)
}
\arguments{
\item{x}{formula}

\item{stringsAsFactors}{cf \code{\link{as.data.frame}}}

\item{\ldots}{unused}
}
\description{
As data frame for data()
}
\examples{
datas <-  data()
head(as.data.frame(datas))
}
