% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxy_datasets.R
\docType{data}
\name{CDL_colors}
\alias{CDL_colors}
\title{CDL crop and color table}
\format{
A data-frame
}
\usage{
CDL_colors
}
\description{
Crop Data Layer table, custom class names
}
\examples{
library(matPkg)
data(CDL_colors)
CDL_colors
}
\keyword{datasets}
