% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\docType{data}
\name{gg_helpers}
\alias{gg_helpers}
\alias{mat_gg_helpers}
\alias{mat_gg_abline_01}
\alias{mat_gg_legend_bottom}
\alias{mat_gg_legend_none}
\alias{mat_gg_title_center}
\alias{mat_gg_scale_Y_toperc}
\alias{mat_gg_scale_X_toperc}
\alias{mat_gg_scale_Y_isperc}
\alias{mat_gg_scale_X_isperc}
\title{ggplot2 helpers}
\format{An object of class \code{LayerInstance} (inherits from \code{Layer}, \code{ggproto}, \code{gg}) of length 11.}
\usage{
mat_gg_abline_01

mat_gg_legend_bottom

mat_gg_legend_none

mat_gg_title_center

mat_gg_scale_Y_toperc

mat_gg_scale_X_toperc

mat_gg_scale_Y_isperc

mat_gg_scale_X_isperc
}
\description{
Gives  a bunch of different helpers
}
\keyword{datasets}
