% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/88_list_pkgs_stuff.R
\name{mat_88_list_pkgs}
\alias{mat_88_list_pkgs}
\alias{mat_88_check_pkgs}
\alias{mat_88_list_paths}
\alias{mat_88_check_paths}
\title{List required packages}
\usage{
mat_88_list_pkgs(scripts_file, warn_missing = TRUE, unique = TRUE)

mat_88_check_pkgs(scripts_file)

mat_88_list_paths(
  scripts_file,
  warn_missing = TRUE,
  unique = TRUE,
  dir_path = "."
)

mat_88_check_paths(scripts_file, dir_path = ".")
}
\arguments{
\item{scripts_file}{data from \code{\link{mat_99_run_Rfiles}}}

\item{warn_missing}{Should (yes) show warning about missing packages?}

\item{unique}{Should (yes) return only distinct packages. Otherwise returns also script name.}

\item{dir_path}{additional directory}
}
\description{
List required packages
}
\examples{
path_rscripts <- system.file("r_scripts_fake", package = "matPkg")
dir_dat <- mat_99_list_Rfiles(path_rscripts)
mat_88_list_pkgs(dir_dat)
mat_88_list_paths(dir_dat)
}
