% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_run_check.R
\name{mat_99_check_there}
\alias{mat_99_check_there}
\title{Write output of 999 file}
\usage{
mat_99_check_there(dir, overwrite = TRUE)
}
\arguments{
\item{dir}{main directory}

\item{overwrite}{should overwtie data?}
}
\description{
Write output of 999 file
}
