% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_run_check.R
\name{mat_99_list_Rfiles}
\alias{mat_99_list_Rfiles}
\alias{mat_list_Rfiles}
\alias{mat_99_run_Rfiles}
\alias{mat_run_Rfiles}
\alias{mat_99_showErr}
\alias{mat_99_check_there_update}
\alias{mat_99_check_there}
\alias{mat_99_write}
\alias{mat_99_add_info_last}
\alias{mat_99_arrange_by_last}
\title{Run-check files 999 machinery}
\usage{
mat_99_list_Rfiles(dir_path, no_old = TRUE, recursive = FALSE)

mat_list_Rfiles(dir_path, no_old = TRUE, recursive = FALSE)

mat_99_run_Rfiles(scripts_file, echo = FALSE)

mat_run_Rfiles(scripts_file, echo = FALSE)

mat_99_showErr(scripts_file_runned)

mat_99_check_there_update(dir_path = "code_setup", overwrite = TRUE)

mat_99_check_there(dir_path, overwrite = TRUE)

mat_99_write(scripts_file_runned, dir_path)

mat_99_add_info_last(scripts_file_runned, dir_path, warn = TRUE)

mat_99_arrange_by_last(scripts_file_runned, dir_path)
}
\arguments{
\item{dir_path}{directory of files}

\item{no_old}{Avoid scripts in directory old?}

\item{recursive}{Look into recursive folders?}

\item{scripts_file}{file of R scripts with path to run, from mat_99_list_Rfiles()}

\item{echo}{print which file done, and gc?}

\item{scripts_file_runned}{data from mat_99_run_Rfiles}

\item{overwrite}{should overwrite data?}

\item{warn}{Should it warn if file is actually missing?}
}
\description{
Run-check files 999 machinery
}
\examples{
library(matPkg)
library(readr)

## read
path_rscripts <- system.file("r_scripts_fake", package = "matPkg")
path_temp <- tempdir()
dir_dat <- mat_99_list_Rfiles(path_rscripts)
dir_dat

## run
out <- mat_99_run_Rfiles(dir_dat)
out
mat_99_showErr(out)

## save
mat_99_write(out, dir = path_temp)

## check output
read_csv(paste(path_temp, "999_CHECK_RUN_report.csv", sep="/"), col_types = cols())

## rerun
out2 <- mat_99_run_Rfiles(dir_dat)
mat_99_write(out2, dir = path_temp)
read_csv(paste(path_temp, "999_CHECK_RUN_report.csv", sep="/"), col_types = cols())

## check format
mat_99_check_there_update(path_temp, overwrite=FALSE)
mat_99_check_there_update(path_temp, overwrite=TRUE)
}
