% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_run_check.R
\name{mat_99_write}
\alias{mat_99_write}
\title{Write output of 999 file}
\usage{
mat_99_write(df, dir)
}
\arguments{
\item{df}{datd from 999}

\item{dir}{Directory}
}
\description{
Write output of 999 file
}
