% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mat_add_perc}
\alias{mat_add_perc}
\title{Add percentage column}
\usage{
mat_add_perc(df, ..., .name = n)
}
\arguments{
\item{df}{data}

\item{.name}{Name of n}

\item{\ldots}{variables to group for}
}
\description{
Add percentage column
}
\examples{
 library(tibble)
 library(magrittr)
 library(dplyr)
  df <- tibble(group = rep(letters[1:2], each = 2),
  n = c(3, 2,3, 5))
  df \%>\% mat_add_perc()
  df \%>\% mat_add_perc(group)
  df \%>\% rename(N=n) \%>\% mat_add_perc(.name = N)
  df \%>\% rename(N=n) \%>\% mat_add_perc(group, .name = N)
}
