% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mat_add_row_num}
\alias{mat_add_row_num}
\title{Add row number}
\usage{
mat_add_row_num(df, ...)
}
\arguments{
\item{df}{data}

\item{\ldots}{group by variables}
}
\description{
Add row number
}
\examples{
df <- data.frame(group = rep(letters[1:2], each=3), values= 1:6)
mat_add_row_num(df)
mat_add_row_num(df, group)
}
