% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mat_add_total_col}
\alias{mat_add_total_col}
\title{Add rowSums over some variables}
\usage{
mat_add_total_col(df, ...)
}
\arguments{
\item{df}{data}

\item{\ldots}{variables}
}
\description{
If not specified, over all numeric columns
}
\examples{
library(dplyr)
df <- data.frame(Char = c("A", "B"),
                 var_1 = c(1,2),
                 var_2=c(1,2),
                 other = c(9,10))
mat_add_total_col(df)
mat_add_total_col(df, var_1)
mat_add_total_col(df, -var_1, -Char)
mat_add_total_col(df, starts_with("var"))
mat_add_total_col(df, starts_with("not found"))
}
