% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_cdl_cols.R
\name{mat_cdl_cols}
\alias{mat_cdl_cols}
\title{Get relevant CDL colors}
\usage{
mat_cdl_cols(CDL_raster)
}
\arguments{
\item{CDL_raster}{a raster containign CDL values}
}
\description{
From a CDL raster, subset the included CDL colors, return them
}
