% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{mat_check_even}
\alias{mat_check_even}
\title{Check number even}
\usage{
mat_check_even(x)
}
\arguments{
\item{x}{vector}
}
\description{
Check number even
}
\examples{
vec <-  1:6
names(vec) <-  vec
mat_check_even(vec)
}
