% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom_tidy.R
\name{mat_coef_wide}
\alias{mat_coef_wide}
\title{Get coef as wide df}
\usage{
mat_coef_wide(reg)
}
\arguments{
\item{reg}{regression object}
}
\description{
Get coef as wide df
}
\examples{
reg <- lm(freeny)
mat_coef_wide(reg)
}
