% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{mat_col_facet}
\alias{mat_col_facet}
\title{Colour facets of ggplot}
\usage{
mat_col_facet(pl, pal = c("#FFD400", "#267000"), side = c("top", "side"))
}
\arguments{
\item{pl}{plot}

\item{pal}{palette}

\item{side}{top or side?}
}
\description{
Colour facets of ggplot
}
\examples{
library(ggplot2)
# facet on side:
pl <- ggplot(aes(x=Sepal.Length, y=Sepal.Width), data = iris) +
  geom_point()+
  facet_grid(Species~.)
plot(mat_col_facet(pl, pal = c("red", "blue", "pink"), side = "side"))

## facet on top:
pl2 <- pl +
 facet_grid(. ~Species)
 plot(mat_col_facet(pl2, pal = c("red", "blue", "pink"), side = "top"))
}
