% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mat_cols_change}
\alias{mat_cols_change}
\title{Change column names}
\usage{
mat_cols_change(df, pattern, replacement)
}
\arguments{
\item{df}{data}

\item{pattern, replacement}{See str_replace}
}
\description{
Change column names
}
\examples{
library(dplyr)
iris \%>\%
  mat_cols_change("Sepal", "pal") \%>\%
  head()
}
