% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_do.R
\name{mat_cor_do}
\alias{mat_cor_do}
\title{compute correlations among all variables}
\usage{
mat_cor_do(df, ...)
}
\arguments{
\item{df}{data}

\item{\ldots}{further argumetns passed to \code{\link{cor.test}}}
}
\description{
compute correlations among all variables
}
\examples{
x <- data.frame(matrix(rnorm(500), ncol=5))
colnames(x) <-  letters[1:5]

mat_cor_do(x)
mat_cor_do(x, method = "kendall")
mat_cor_do(x, method = "kendall", alternative = "greater")
}
