% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{mat_cor_na}
\alias{mat_cor_na}
\alias{mat_cor_na_df}
\title{Cor with NA}
\usage{
mat_cor_na(x, y = NULL, use = "pairwise.complete.obs", method = "pearson")

mat_cor_na_df(x, y = NULL, use = "pairwise.complete.obs", method = "pearson")
}
\arguments{
\item{x, y, use, method}{as in cor}
}
\description{
Cor with NA
}
\examples{
iris2 <- iris[, 1:4]
iris2[c(3, 8),1] <-  NA
mat_cor_na(iris2)
with(iris2, mat_cor_na(Sepal.Length, Petal.Width))
with(iris2, mat_cor_na_df(Sepal.Length, Petal.Width))
}
