% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.R
\name{mat_cor_to_df}
\alias{mat_cor_to_df}
\title{Convert correlation object to df}
\usage{
mat_cor_to_df(x, long = TRUE)
}
\arguments{
\item{x}{object from cor}

\item{long}{into long format?}
}
\description{
Convert correlation object to df
}
\examples{
library(magrittr)
library(dplyr)
data(iris_tb)
iris_tb \%>\%
  select(-Species) \%>\%
  cor()  \%>\%
  mat_cor_to_df()
}
