% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_separate.R
\name{mat_count_separate}
\alias{mat_count_separate}
\title{Count each column separately}
\usage{
mat_count_separate(df, ...)
}
\arguments{
\item{df}{initial data}

\item{\ldots}{tidy selectors}
}
\description{
Count each column separately
}
\examples{
df <- data.frame(ids = 1:100,
                 x1=sample(letters[1:5], 100, replace = TRUE),
                 x5=sample(c("TRUE", "FALSE"), 100, replace = TRUE))
mat_count_separate(df, starts_with("x"))
}
