% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_helper.R
\name{mat_cross_helper}
\alias{mat_cross_helper}
\title{Crossing with helpers}
\usage{
mat_cross_helper(df, args, wide = FALSE)
}
\arguments{
\item{df}{data}

\item{args}{a list, names, with values. Special are .each, and .all}

\item{wide}{to wide?}
}
\description{
Crossing with helpers
}
\examples{
mat_cross_helper(warpbreaks, list(wool=c("A"),
                           tension = c("L", "H")))
mat_cross_helper(warpbreaks, list(wool=c("A"),
                           tension = c(".each")))
}
