% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mat_df_trim_quant}
\alias{mat_df_trim_quant}
\title{Trim a df by quantiles}
\usage{
mat_df_trim_quant(
  df,
  .value_var,
  ...,
  .probs = c(0.02, 0.98),
  na.rm = TRUE,
  .rem_quants = TRUE
)
}
\arguments{
\item{df}{The df}

\item{.value_var}{the column of values}

\item{.probs}{quantile probs}

\item{na.rm}{Remove NA when computing quantile?}

\item{.rem_quants}{remove the quantile (i.e. trim only)?}

\item{\ldots}{the group_by variables}
}
\description{
Trim a df by quantiles
}
\examples{
mat_df_trim_quant(iris, Petal.Width, Species)
}
