% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mat_enframe_wide}
\alias{mat_enframe_wide}
\title{enframe wide}
\usage{
mat_enframe_wide(x, names = NULL)
}
\arguments{
\item{x}{a vector}

\item{names}{vector of names}
}
\description{
Add a col diff
}
\examples{
mat_enframe_wide(x=c(a=1,b=22))
mat_enframe_wide(x=c(1,22))
mat_enframe_wide(x=c(a=1,22, c=9, 8))
}
