% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{mat_find_method}
\alias{mat_find_method}
\title{Show method called on object}
\usage{
mat_find_method(generic, ...)
}
\arguments{
\item{generic}{the generic}

\item{\ldots}{the object on which to use the generic}
}
\description{
Show method called on object
}
\examples{
data(iris_tb)
mat_find_method(print, iris)
mat_find_method(print, iris_tb)
}
