% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_various.R
\name{mat_formu_to_char}
\alias{mat_formu_to_char}
\title{Formula to character}
\usage{
mat_formu_to_char(x, ...)
}
\arguments{
\item{x}{formula}

\item{\ldots}{unused}
}
\description{
Formula to character
}
\examples{
f <-  aha ~ hih + oho
as.character(f)
}
