% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{mat_geom_lm}
\alias{mat_geom_lm}
\title{geom_lm: enforce lm without se}
\usage{
mat_geom_lm(
  mapping = NULL,
  data = NULL,
  stat = "smooth",
  position = "identity",
  ...,
  method = "lm",
  formula = y ~ x,
  se = FALSE,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping, data, stat, position}{Unchanged}

\item{...}{As in geom_smooth}

\item{method, se}{Arguments changed}

\item{formula, na.rm, show.legend, inherit.aes}{Unchanged}
}
\description{
geom_lm: enforce lm without se
}
\examples{
library(ggplot2)
ggplot(mpg, aes(displ, hwy)) +
  geom_point() +
  mat_geom_lm()
}
