% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_group.R
\name{mat_group_quantile}
\alias{mat_group_quantile}
\title{Compute quantiles by group}
\usage{
mat_group_quantile(df, .value_var, ..., .probs = c(0.05, 0.25, 0.5, 0.75,
  0.95), na.rm = TRUE)
}
\arguments{
\item{df}{The df}

\item{.value_var}{the column of values}

\item{.probs}{quantile probs}

\item{na.rm}{Remove NA when computing quantile?}

\item{\ldots}{the group_by variables}
}
\description{
Compute quantiles by group
}
\examples{
mat_group_quantile(iris, Petal.Width, Species)
}
\seealso{
\code{\link{mat_df_trim_quant}}
}
