% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mat_head_group}
\alias{mat_head_group}
\title{Take first groups}
\usage{
mat_head_group(df, ..., n_head = 5)
}
\arguments{
\item{df}{data}

\item{...}{grouping vars}

\item{n_head}{Number of groups}
}
\description{
Take first groups
}
\examples{
df <- data.frame(group_a = rep(LETTERS[1:3], each=4),
                 group_b = rep(letters[1:2], 6),dat = 1:12)
mat_head_group(df, group_a, n_head=1)
mat_head_group(df, group_a, group_b, n_head=1)
}
