% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mat_join_compare}
\alias{mat_join_compare}
\title{Compare if datasets are equal}
\usage{
mat_join_compare(
  df1,
  df2,
  by = NULL,
  join_fun = dplyr::inner_join,
  tol = 1e-08
)
}
\arguments{
\item{df1, df2}{The two df to compare}

\item{by}{the key variables}

\item{join_fun}{the type of join}

\item{tol}{tolerance value}
}
\description{
Merge, and for all identical variables, compare
}
\examples{
library(dplyr)
data(iris_tb)
iris_orig <- mutate(iris_tb, row_num = 1:n())
iris_new <- mutate(iris_orig, Sepal.Length =Sepal.Length+0.000001)
mat_join_compare(df1=iris_orig, df2=iris_new, by = c("row_num", "Species"))
mat_join_compare(df1=iris_orig, df2=iris_new, by =c("row_num", "Species"), tol = 0.00001)
}
