% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{mat_keep_first}
\alias{mat_keep_first}
\title{remove duplicated in vector}
\usage{
mat_keep_first(x, fill = "", warn = TRUE)
}
\arguments{
\item{x}{vector}

\item{fill}{which value to use}

\item{warn}{SHould warn for non-sorted?}
}
\description{
remove duplicated in vector
}
\examples{
mat_keep_first(rep(1:3, 1:3))
}
