% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mat_li_comp_cols}
\alias{mat_li_comp_cols}
\title{Compare list of names}
\usage{
mat_li_comp_cols(list_df, logi = FALSE)
}
\arguments{
\item{list_df}{List of df}

\item{logi}{return TRUE instead of class}
}
\description{
Compare list of names
}
\examples{
freeny_2 <-  dplyr::rename(freeny, price.index2 = price.index)
li <- list(freeny, freeny_2)
mat_li_comp_cols(list =li)
}
