% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_directories.R
\name{mat_list_dir}
\alias{mat_list_dir}
\title{List files in a directory}
\usage{
mat_list_dir(
  path,
  pattern = "\\\\.tif",
  recursive = TRUE,
  add_ext = FALSE,
  ...
)
}
\arguments{
\item{path}{dir path}

\item{pattern, recursive}{passed to list.files}

\item{add_ext}{Add column with extension?}

\item{\ldots}{Additional arguments to \code{\link{list.files}}}
}
\description{
List files in a directory
}
\examples{
mat_list_dir(path = ".", pattern = "R")
}
