% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_run_check.R
\name{mat_list_Rfiles}
\alias{mat_list_Rfiles}
\title{List yaml of R scripts}
\usage{
mat_list_Rfiles(dir_path)
}
\arguments{
\item{dir_path}{directory of files}
}
\description{
List yaml of R scripts
}
\examples{
mat_list_Rfiles("R")
}
