% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom_tidy.R
\name{mat_lm_means_tidy}
\alias{mat_lm_means_tidy}
\title{Compute means by lm, tidy, unnest}
\usage{
mat_lm_means_tidy(df_nest, val_name = value, clean = TRUE,
  vcov. = NULL)
}
\arguments{
\item{df_nest}{a nested df}

\item{val_name}{name of value column}

\item{clean}{Clean the broom colnames?}

\item{vcov.}{Specify vcov}
}
\description{
This returns a clean df of means for each group
}
\examples{
library(magrittr)
data(iris_tb)
iris_tb \%>\%
  tidyr::nest(-Species) \%>\%
  mat_lm_means_tidy(Petal.Width)
}
