% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_run_check.R
\name{mat_parse_yaml}
\alias{mat_parse_yaml}
\title{Read YAML header to list}
\usage{
mat_parse_yaml(file_path, encoding = getOption("encoding"))
}
\arguments{
\item{file_path}{The file path}

\item{encoding}{just in case}
}
\description{
Read YAML header to list
}
