% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_to_pdf.R
\name{mat_pdf_to_png}
\alias{mat_pdf_to_png}
\title{Convert pdf to png}
\usage{
mat_pdf_to_png(path, density = 150, quality = 90, echo = TRUE)
}
\arguments{
\item{path}{path of file}

\item{density, quality}{parameters to convert}

\item{echo}{should show command?}
}
\description{
Convert pdf to png
}
\examples{
  pdf("try.pdf")
  plot(1, col = "green")
  dev.off()
  mat_pdf_to_png("try.pdf")
  file.remove(c("try.pdf", "try.png"))
}
