% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coefs_tidy.R
\name{mat_plot_coefs_tidy}
\alias{mat_plot_coefs_tidy}
\title{Plot coefficients from tidy output}
\usage{
mat_plot_coefs_tidy(df, fill_var = term, fac1_var, fac2_var, angle = 0,
  x_var = term, scales = "fixed")
}
\arguments{
\item{df}{Data-frame}

\item{fill_var}{variable for \code{fill}}

\item{fac1_var, fac2_var}{facet variable}

\item{angle}{angle in \code{element_text()}}

\item{x_var}{x variable, default term}

\item{scales}{argument to \code{facet_grid}}
}
\description{
Plot coefficients from tidy output
}
\examples{
data(coefs_out_iris)
mat_plot_coefs_tidy(coefs_out_iris, fac1_var = Species)
mat_plot_coefs_tidy(coefs_out_iris, fac2_var = Species)
mat_plot_coefs_tidy(coefs_out_iris, fac1_var = Species, fac2_var = term, scales = "free")

## handle duplicates
mat_plot_coefs_tidy(coefs_out_iris)
}
