% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{mat_print_tib}
\alias{mat_print_tib}
\title{Set options for tibble printing}
\usage{
mat_print_tib(min = 3, max = 6)
}
\arguments{
\item{min}{Arg to \code{pillar.print_min = min}}

\item{max}{Arg to \code{pillar.print_min = max}}
}
\description{
Set options for tibble printing
}
\examples{
mat_print_tib(2, 4)
as_tibble(iris) \%>\% head(3)
as_tibble(iris) \%>\% head(4)
as_tibble(iris) \%>\% head(5)
}
