% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{mat_pval_cat}
\alias{mat_pval_cat}
\title{Add p-value and paste}
\usage{
mat_pval_cat(estimate, p_value, digits = 2, ...)
}
\arguments{
\item{estimate}{Parameters of interest}

\item{p_value}{p-values for the estimate}

\item{digits}{rounding before pasting}

\item{\ldots}{Currently unused}
}
\description{
Add p-value and paste
}
\examples{
p_vals <- c(0.01, 0.0001, 0.05, 0.09)
mat_pval_cat(p_vals, p_vals)
mat_pval_cat(p_vals, p_vals, digits=4)
}
\seealso{
\code{\link{format.pval}}
}
