% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.R
\name{mat_quant_to_df}
\alias{mat_quant_to_df}
\alias{quant_to_df}
\title{Convert quantile() output to df}
\usage{
mat_quant_to_df(x, long = FALSE)

quant_to_df(x)
}
\arguments{
\item{x}{output from quantile}

\item{long}{Use long format?}
}
\description{
Convert quantile() output to df
}
\examples{
X <- runif(100)
q <- quantile(X, probs= c(0.001, 0.975))
mat_quant_to_df(q)
mat_quant_to_df(q, long=TRUE)
}
