% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mat_show_cols_1val}
\alias{mat_show_cols_1val}
\alias{mat_remo_cols_1val}
\title{Remove columns with only one value}
\usage{
mat_show_cols_1val(df, nval_max = 1, wide = TRUE, show_na = TRUE)

mat_remo_cols_1val(df, ...)
}
\arguments{
\item{df}{data}

\item{nval_max}{how many values allowed?}

\item{wide}{logical TRUE, if nval_max > 1, show in wide format?}

\item{show_na}{default TRUE. Should show all NA?}

\item{\ldots}{variables to keep}
}
\description{
Remove columns with only one value
}
\examples{
data(quick_stats)
mat_show_cols_1val(quick_stats)
mat_show_cols_1val(quick_stats, nval_max = 2)
mat_show_cols_1val(quick_stats, nval_max = 5, show_na = FALSE)

## now remove
mat_remo_cols_1val(quick_stats)
mat_remo_cols_1val(quick_stats, geo_level)
}
