% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_run_check.R
\name{mat_run_Rfiles}
\alias{mat_run_Rfiles}
\title{Run files from a df-df}
\usage{
mat_run_Rfiles(df)
}
\arguments{
\item{df}{the df of files, from mat_list_Rfiles()}
}
\description{
Run files from a df-df
}
