% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safely.R
\name{mat_safely_unnest}
\alias{mat_safely_unnest}
\title{Unest safely}
\usage{
mat_safely_unnest(df, col_name, result_name = "result")
}
\arguments{
\item{df}{data}

\item{col_name}{which column contains the safely output?}

\item{result_name}{How to call output column with result?}
}
\description{
Unest safely
}
\examples{
library(purrr)
library(dplyr)
df <- data.frame(a=c(1, NA, -2))
df_out <- mutate(df, out= map(a, ~safely(~log2(., arg = FALSE))(.)))
mat_safely_unnest(df_out, out)

}
