% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memeory_show_check.R
\name{mat_show_mem}
\alias{mat_show_mem}
\title{Show memory usage, by object}
\usage{
mat_show_mem(df_input)
}
\arguments{
\item{df_input}{input df. If empty, will evaluate ls()}
}
\description{
Show memory usage, by object
}
\examples{
mat_show_mem()
mat_show_mem(freeny)
library(purrr)
library(dplyr)
library(tidyr)

iris_regs <- nest(iris, data=-Species) \%>\%
  mutate(reg_out = map(data, ~lm(Petal.Width~Petal.Length, data=as_tibble(.)))) \%>\%
  select(-data)
mat_show_mem(iris_regs$reg_out)
}
