% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mat_slice_by}
\alias{mat_slice_by}
\title{Slice by group}
\usage{
mat_slice_by(df, N = 100, ...)
}
\arguments{
\item{df}{data}

\item{N}{(max) number per group}

\item{\ldots}{group variables}
}
\description{
Slice by group
}
\examples{
mat_slice_by(iris, N = 10, Species)
}
