% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mat_spread_TR_FALSE}
\alias{mat_spread_TR_FALSE}
\title{Spread data with TRUE/FALSE (from count)}
\usage{
mat_spread_TR_FALSE(df, col, n_col = n)
}
\arguments{
\item{df}{data}

\item{col}{col to spread}

\item{n_col}{name of n column}
}
\description{
Spread data with TRUE/FALSE (from count)
}
