% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_functions.R
\name{mat_st_df_to_pt}
\alias{mat_st_df_to_pt}
\title{mat_st_to_df}
\usage{
mat_st_df_to_pt(df, lat_var, lng_var, crs = NULL)
}
\arguments{
\item{df}{data frame}

\item{lat_var, lng_var}{name of cloumns}

\item{crs}{crs}
}
\description{
mat_st_to_df
}
\examples{
df <- data.frame(x = c(1,2,3), y = c(3,4,2))
df_st <- mat_st_df_to_pt(df, lat_var = x, lng_var = y, crs = 4326)
plot(df_st)
}
