% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_functions.R
\name{mat_st_read}
\alias{mat_st_read}
\title{Read st silently}
\usage{
mat_st_read(dsn, ..., quiet = TRUE, stringsAsFactors = FALSE)
}
\arguments{
\item{dsn, quiet, stringsAsFactors}{check doc}

\item{\ldots}{passed too}
}
\description{
Simpelr fun
}
