% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mat_str_to_expr}
\alias{mat_str_to_expr}
\title{Converting to expression for plots}
\usage{
mat_str_to_expr(x)
}
\arguments{
\item{x}{A single charracter string}
}
\value{
An expression
}
\description{
Simply add annoying ~ as space, and converts as expression
}
\examples{
plot(1)
title(mat_str_to_expr("This is R^2"))
}
