% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_to_pdf.R
\name{mat_table_to_pdf}
\alias{mat_table_to_pdf}
\alias{table_to_pdf}
\title{Convert a table to direct pdf}
\usage{
mat_table_to_pdf(
  x,
  filename = "input.pdf",
  quiet = TRUE,
  is_path_x = FALSE,
  clean_tex = TRUE,
  clean_rest = TRUE,
 
    plus = "\\\\usepackage{booktabs}\\n\\\\usepackage{dcolumn}\\n\\\\usepackage{underscore}"
)

table_to_pdf(x)
}
\arguments{
\item{x}{the table input}

\item{filename}{output name}

\item{quiet}{passed to tools::texi2pdf}

\item{is_path_x}{whether x is a file on disk}

\item{clean_tex, clean_rest}{Should remove other files, and tex too?}

\item{plus}{additional lattex stuff}
}
\description{
Use tools::texi2pdf
}
\examples{
\dontrun{
tab <-  data.frame(variable= c("A", "B"), stat = c("mean"), value = 1:2)
tab_xt <- print(xtable::xtable(tab), print.results = FALSE)

mat_table_to_pdf(tab_xt, filename = "filename_test.pdf")
file.remove("filename_test.pdf")
## case 2: file is already on disk, use is_path_x
tmp_file <- tempfile()
print(xtable::xtable(tab), file = tmp_file)
mat_table_to_pdf(tmp_file, filename = "filename_test2.pdf", is_path_x = TRUE)
file.remove("filename_test2.pdf")
}
}
\seealso{
\code{\link{mat_pdf_to_png}}
}
