% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom_tidy.R
\name{mat_tidy_clean}
\alias{mat_tidy_clean}
\title{Tidy columns: . to _}
\usage{
mat_tidy_clean(df)
}
\arguments{
\item{df}{data}
}
\description{
Tidy columns: . to _
}
\examples{
reg <- lm(freeny)
out_1 <- broom::tidy(reg)
out_2 <- broom::glance(reg)
mat_tidy_clean(out_1)
mat_tidy_clean(out_2)
}
