% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom_tidy.R
\name{mat_tidy_coef_plot}
\alias{mat_tidy_coef_plot}
\title{For broom coef tidy output: bar plot for coefs}
\usage{
mat_tidy_coef_plot(df, x_var, fill_var, fac_1 = NULL, fac_2 = NULL)
}
\arguments{
\item{df}{data}

\item{x_var}{var used in x axis (y is estimate)}

\item{fill_var}{var used to have multiple x}

\item{fac_1, fac_2}{vars for facetting}
}
\description{
For broom coef tidy output: bar plot for coefs
}
