% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom_tidy.R
\name{mat_tidy_do}
\alias{mat_tidy_do}
\title{Tidy on a df of regs with col reg_out}
\usage{
mat_tidy_do(df, reg_col = reg_out, ...)
}
\arguments{
\item{df}{data}

\item{reg_col}{name of column with regs?}

\item{\ldots}{futher arguments passed to tidy}
}
\description{
Tidy on a df of regs with col reg_out
}
\examples{
library(purrr)
library(dplyr)
library(tidyr)

iris_regs <- nest(iris, data=-Species) \%>\%
mutate(reg_out = map(data, ~lm(Petal.Width~Petal.Length, data=as_tibble(.)))) \%>\%
select(-data)

mat_tidy_do(df=iris_regs)
}
