% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom_tidy.R
\name{mat_tidy_glance}
\alias{mat_tidy_glance}
\title{Tidy and glance together}
\usage{
mat_tidy_glance(reg, wide = TRUE)
}
\arguments{
\item{reg}{regression object}

\item{wide}{Do wide (removing pval, se, etc...)}
}
\description{
Tidy and glance together
}
\examples{

reg <- lm(freeny)
mat_tidy_glance(reg)
mat_tidy_glance(reg, wide=FALSE)
}
