% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom_tidy.R
\name{mat_tidy_robust}
\alias{mat_tidy_robust}
\title{Use tidy() with robust vcov directly}
\usage{
mat_tidy_robust(
  x,
  vcov. = NULL,
  conf.int = TRUE,
  df = NULL,
  parm = NULL,
  level = 0.95,
  ...
)
}
\arguments{
\item{x}{object for which hopfully there is a coeftst/vcov method}

\item{conf.int}{return confindence interval? Determined by level}

\item{df, parm, level, vcov.}{arguments to \code{\link{coeftest}}}

\item{\ldots}{passed to vcov (well to coeftest who does the jo)}
}
\description{
Use tidy() with robust vcov directly
}
\examples{
library(sandwich)
reg <-  lm(freeny)
standard_me <- mat_tidy_robust(reg)
standard_br <- broom::tidy(reg, conf.int = TRUE)
all.equal(as.data.frame(standard_me), as.data.frame(standard_br))
mat_tidy_robust(reg, vcov. = vcovHC(reg))
mat_tidy_robust(reg, vcov. = vcovHC)
}
