% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mat_vars_uniques}
\alias{mat_vars_uniques}
\title{Table of values of variables}
\usage{
mat_vars_uniques(df, wide = TRUE, ...)
}
\arguments{
\item{df}{data-frame}

\item{wide}{Wide the dataset?}

\item{\ldots}{variables to keep}
}
\description{
Table of values of variables
}
\examples{
mat_vars_uniques(iris)
data_test <- tidyr::crossing(a=letters[1:3], b=c("blue", "green"))
mat_vars_uniques(data_test)
mat_vars_uniques(df=data_test, wide=FALSE)
}
