% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matPkg.R
\docType{package}
\name{matPkg}
\alias{matPkg}
\title{matPkg: Internal functions I use}
\description{
A personal package for functions I use often. Use at your own risk, no guarantee of being consistent.
}
\details{
Should run:

remotes::install_github("MatthieuStigler/matPkg", upgrade = "never")
}
