% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.table.R
\name{setTb}
\alias{setTb}
\title{Set (back) to tibble}
\usage{
setTb(df, value = c("tbl_df", "tbl", "data.frame"))
}
\arguments{
\item{df}{initial data}

\item{value}{The old class}
}
\description{
Set (back) to tibble
}
\examples{
library(data.table)
iris_test <- copy(tibble::as_tibble(iris))
setDT(iris_test)
iris_test
setTb(iris_test)
iris_test
}
