% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_functions.R
\name{st_as_sf.Extent}
\alias{st_as_sf.Extent}
\title{Extent to sf}
\usage{
st_as_sf.Extent(x)
}
\arguments{
\item{x}{raster Extent object}
}
\description{
Extent to sf
}
\examples{
library(raster)
library(sf)
r1 <- raster(nrows=108, ncols=21, xmn=0, xmx=100)
r2 <- raster(nrows=108, ncols=21, xmn=50, xmx=150)
st2 <- rbind(st_as_sf(extent(r1)),
             st_as_sf(extent(r2)))
             plot(st2, border = 1:2)

}
