% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/98_list_pkgs_stuff.R
\name{mat_88_list_pkgs}
\alias{mat_88_list_pkgs}
\title{List required packages}
\usage{
mat_88_list_pkgs(scripts_file, warn_missing = TRUE, unique = TRUE)
}
\arguments{
\item{scripts_file}{data from \code{\link{mat_99_run_Rfiles}}}

\item{warn_missing}{Should (yes) show warning about missing packages?}

\item{unique}{Should (yes) return only distinct packages. Otherwise returns also script name.}
}
\description{
List required packages
}
\examples{
path_rscripts <- system.file("r_scripts_fake", package = "matPkg")
dir_dat <- mat_99_list_Rfiles(path_rscripts)
mat_88_list_pkgs(dir_dat)
}
