% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mat_add_perc}
\alias{mat_add_perc}
\title{Add percentage column}
\usage{
mat_add_perc(x, ..., .name = n)
}
\description{
Add percentage column
}
\examples{
  df <- tibble(group = rep(letters[1:2], each = 2),
  n = c(3, 2,3, 5))
  df \%>\% add_perc()
  df \%>\% add_perc(group)
  df \%>\% rename(N=n) \%>\% add_perc(.name = N)
  df \%>\% rename(N=n) \%>\% add_perc(group, .name = N)
}
