% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{mat_check_0row}
\alias{mat_check_0row}
\title{Check 0 rows}
\usage{
mat_check_0row(df, message_ok = "OK")
}
\arguments{
\item{df}{data}

\item{message_ok}{the ok message}
}
\description{
Check 0 rows
}
\examples{
mat_check_0row(df=subset(iris, Species=="aaa"))
}
