% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mat_col_check_same}
\alias{mat_col_check_same}
\title{Compare if col_1 and col_2 equal}
\usage{
mat_col_check_same(df, col_1, col_2, tol = 1e-08)
}
\arguments{
\item{df}{data}

\item{col_1, col_2}{The columns}

\item{tol}{tolerance level}
}
\description{
Add a col diff
}
\examples{
library(dplyr)
data(iris_tb)
mutate(iris_tb, Sepal.Length2 =Sepal.Length+rep(c(0.000001, 0), 75)) \%>\%
  mat_col_check_same(Sepal.Length2, Sepal.Length)
}
