% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mat_is_unique_combo}
\alias{mat_is_unique_combo}
\title{Trim a df by quantiles}
\usage{
mat_is_unique_combo(df, ..., .print = TRUE)
}
\arguments{
\item{df}{The df}

\item{.print}{Print bad output?}

\item{\ldots}{the group_by variables}
}
\description{
Trim a df by quantiles
}
\examples{
mat_is_unique_combo(iris, Sepal.Length, Sepal.Width, Petal.Width, Petal.Length, Species)
}
