% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mat_one_of_quiet}
\alias{mat_one_of_quiet}
\title{Use one_of() quietly}
\usage{
mat_one_of_quiet(df)
}
\arguments{
\item{df}{data}
}
\description{
Use one_of() quietly
}
\examples{
library(magrittr)
library(tibble)
library(dplyr)
iris \%>\% as_tibble() \%>\%
select(mat_one_of_quiet(c("Sepal.Length", "caca")))
}
