% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_to_pdf.R
\name{mat_pdf_to_png}
\alias{mat_pdf_to_png}
\title{Convert pdf to png}
\usage{
mat_pdf_to_png(
  path,
  density = 150,
  quality = 90,
  echo = TRUE,
  correct_grayscale = FALSE
)
}
\arguments{
\item{path}{path of file}

\item{density, quality}{parameters to convert}

\item{echo}{should show command?}

\item{correct_grayscale}{Should correct for \code{RGB color space not permitted on grayscale}}
}
\description{
Convert pdf to png
}
\examples{
tmp <- file.path(tempdir(), "try.pdf")
pdf(tmp)
plot(1, col = "green")
dev.off()
mat_pdf_to_png(tmp)

## use `correct_grayscale = TRUE` for message "RGB color space not permitted on grayscale"
pdf(tmp)
plot(1)
dev.off()
mat_pdf_to_png(tmp, correct_grayscale = TRUE)
}
\seealso{
\code{\link{mat_table_to_pdf}}, \code{\link{mat_tex_to_png}}
}
