% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mat_show_cols_1val}
\alias{mat_show_cols_1val}
\alias{mat_remo_cols_1val}
\title{Remove columns with only one value}
\usage{
mat_show_cols_1val(df, nval_max = 1)

mat_remo_cols_1val(df, ...)
}
\arguments{
\item{df}{data}

\item{nval_max}{how many values allowed?}

\item{\ldots}{variables to keep}
}
\description{
Remove columns with only one value
}
\examples{
data(quick_stats)
mat_show_cols_1val(quick_stats)
mat_remo_cols_1val(quick_stats)
mat_remo_cols_1val(quick_stats, geo_level)
}
