% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_run_check.R
\name{mat_run_Rfiles}
\alias{mat_run_Rfiles}
\title{Run files from a df-df}
\usage{
mat_run_Rfiles(df, echo = FALSE)
}
\arguments{
\item{df}{the df of files, from mat_list_Rfiles()}

\item{echo}{print shich file done, and gc?}
}
\description{
Run files from a df-df
}
\examples{
files_df <- mat_list_Rfiles(system.file("r_scripts_fake", package = "matPkg"))
mat_run_Rfiles(files_df, echo=TRUE)
}
