% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_to_pdf.R
\name{mat_table_to_pdf}
\alias{mat_table_to_pdf}
\title{Convert a table to direct pdf}
\usage{
mat_table_to_pdf(x, filename = "input.tex", clean = TRUE,
  quiet = TRUE,
  plus = "\\\\usepackage{booktabs}\\n\\\\usepackage{dcolumn}\\n\\\\usepackage{underscore}")
}
\description{
Convert a table to direct pdf
}
