% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_to_pdf.R
\name{mat_table_to_pdf}
\alias{mat_table_to_pdf}
\alias{table_to_pdf}
\title{Convert a table to direct pdf}
\usage{
mat_table_to_pdf(x, filename = "input.tex", clean = TRUE,
  quiet = TRUE,
  plus = "\\\\usepackage{booktabs}\\n\\\\usepackage{dcolumn}\\n\\\\usepackage{underscore}")

table_to_pdf(x)
}
\arguments{
\item{x}{the table input}

\item{filename}{output name}

\item{clean, quiet}{passed to tools::texi2pdf}

\item{plus}{additional lattex stuff}
}
\description{
Use tools::texi2pdf
}
