% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_to_pdf.R
\name{mat_tex_to_png}
\alias{mat_tex_to_png}
\title{\subsection{tex to pdf}{

Convert tex to png
}}
\usage{
mat_tex_to_png(
  path,
  path_out = gsub("\\\\.tex$", ".png", path),
  correct_grayscale = TRUE,
  step1_args = list(),
  step2_args = list()
)
}
\arguments{
\item{path}{Path for the .tex on disk}

\item{path_out}{Where to save .png}

\item{correct_grayscale}{Arguments passed to \code{\link{mat_pdf_to_png}}}

\item{step1_args, step2_args}{Lists of arguments passed to  \code{\link{mat_table_to_pdf}} and \code{\link{mat_pdf_to_png}}.}
}
\description{
This is a wrapper for mat_table_to_pdf and mat_pdf_to_png
}
\examples{
\dontrun{
tab <-  data.frame(variable= c("A", "B"), stat = c("mean"), value = 1:2)
tab_xt <- print(xtable::xtable(tab), print.results = FALSE)

##
tmp_file <- tempfile()
print(xtable::xtable(tab), file = tmp_file)
mat_tex_to_png(tmp_file, "table_as_image.png")

## Pass other argumetns to mat_table_to_pdf or mat_pdf_to_png
mat_tex_to_png(tmp_file, "table_as_image.png", step1_args = list(quiet=FALSE))
mat_tex_to_png(tmp_file, "table_as_image.png", step2_args = list(echo=FALSE))

## remove file
file.remove("table_as_image")
}
}
\seealso{
\code{\link{mat_table_to_pdf}}, \code{\link{mat_pdf_to_png}}
}
