% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matPkg-package.R
\docType{package}
\name{matPkg-package}
\alias{matPkg}
\alias{matPkg-package}
\title{A personal package for functions I use often. Use at your own risk, no guarantee of being consistent.}
\description{
Should run:
}
\details{
remotes::install_github("MatthieuStigler/matPkg", upgrade = "never")
}
\author{
\strong{Maintainer}: Matthieu Last \email{Matthieu.Stigler@gmail.com} [copyright holder]

}
\keyword{internal}
